requirements.txt
aiosqlite==0.21.0
APScheduler==3.11.0
discord.py==2.6.3
filelock==3.19.1
Unidecode==1.4.0
yt_dlp==2025.9.23
PyNaCl==1.5.0
flask==3.1.2
gtts==2.5.4
samtts==0.2.0
python-dotenv==1.2.1